Added server common client modules

+++ common/src/main/java/com/haloandrei/socket/common/ServiceMovieInterface.java
@@ -14,7 +14,7 @@ public interface ServiceMovieInterface {
 
     void addMovie(Movie movie) throws MovieRentalException;
 
-    void deleteMovie(Integer id) throws MovieRentalException;
+    void deleteMovie(Long id) throws MovieRentalException;
 
     void updateMovie(Movie movie) throws MovieRentalException;
